refactor: clarify module name
