+++ packages_rs/nextclade/src/coord/mod.rs
@@ -1,4 +1,4 @@
 pub mod coord_map;
-pub mod coord_map2;
+pub mod coord_map_cds_to_global;
 pub mod coord_map_global;
 pub mod coord_map_local;
